Chor: create Extension Character, String
- Character, String Extenstion 생성하여 분리
