+++ Sources/KoTextCountLimit/Extension+Character.swift
@@ -1,0 +1,14 @@
+//
+//  File.swift
+//  
+//
+//  Created by Chung Wussup on 9/12/24.
+//
+
+import Foundation
+
+extension Character {
+    var isConsonant: Bool {
+        return String(self).isConsonant
+    }
+}
